Split stats_manager and daily_stats_reporter
 This is to allow multiple reporters that report data in regular intervals in the future
