+++ config-example.yaml
@@ -27,9 +27,10 @@ keep_alive_monitor:
 
 # Enable this and you'll receive a daily summary notification
 # on your farm performance at the specified time of the day.
-daily_stats:
-  enable: true
-  time_of_day: 21
+stats:
+  daily_stats:
+    enable: true
+    time_of_day: 21
 
 # We support a lot of notifiers, please check the README for more
 # information. You can delete the sections which you aren't using.
