+++ src/chia_log/handlers/daily_stats/stats_manager.py
@@ -1,91 +1,61 @@
 # std
 import logging
-from datetime import datetime, timedelta
-from typing import List
 from threading import Thread
 from time import sleep
+from typing import List
 
 # project
-from . import HarvesterActivityConsumer, FinishedSignageConsumer
-from .stat_accumulators.eligible_plots_stats import EligiblePlotsStats
-from .stat_accumulators.search_time_stats import SearchTimeStats
-from .stat_accumulators.signage_point_stats import SignagePointStats
-from .stat_accumulators.found_proof_stats import FoundProofStats
-from .stat_accumulators.number_plots_stats import NumberPlotsStats
-from src.chia_log.parsers.harvester_activity_parser import HarvesterActivityMessage
+from src.chia_log.handlers.daily_stats import StatReporter
+from src.chia_log.handlers.daily_stats.stat_reporters.daily_stats_reporter import DailyStatsReporter
 from src.chia_log.parsers.finished_signage_point_parser import FinishedSignagePointMessage
+from src.chia_log.parsers.harvester_activity_parser import HarvesterActivityMessage
+from src.config import Config
 from src.notifier.notify_manager import NotifyManager
-from src.notifier import Event, EventType, EventPriority, EventService
 
 
 class StatsManager:
-    """Manage all stat accumulators and trigger daily notification to the user
-    with a summary from all stats that have been collected for the past 24 hours.
+    """Initialize and manage all stat reporters and trigger their reporting.
     """
 
-    def __init__(self, config: dict, notify_manager: NotifyManager):
-        try:
-            self._enable = config["enable"]
-            self._time_of_day = config["time_of_day"]
-        except KeyError as key:
-            logging.error(f"Invalid config.yaml. Missing key: {key}")
-            self._enable = False
-
-        if not self._enable:
-            logging.warning("Disabled stats and daily notifications")
-            return
-
-        logging.info("Enabled stats for daily notifications")
+    def __init__(self, config: Config, notify_manager: NotifyManager):
+        self._config = config.get_stats_config()
         self._notify_manager = notify_manager
-        self._stat_accumulators = [
-            FoundProofStats(),
-            SearchTimeStats(),
-            NumberPlotsStats(),
-            EligiblePlotsStats(),
-            SignagePointStats(),
-        ]
+        self._stats_reporter: dict[str, StatReporter] = {}
+        self._initialize_stats_managers()
 
-        logging.info(f"Summary notifications will be sent out daily at {self._time_of_day} o'clock")
-        self._datetime_next_summary = datetime.now().replace(hour=self._time_of_day, minute=0, second=0, microsecond=0)
-        if datetime.now() > self._datetime_next_summary:
-            self._datetime_next_summary += timedelta(days=1)
+    def _initialize_stats_managers(self):
+        key_notifier_mapping = {
+            "daily_stats": DailyStatsReporter
+        }
 
-        # Start thread
-        self._is_running = True
-        self._thread = Thread(target=self._run_loop)
-        self._thread.start()
+        for key in self._config.keys():
+            if key not in key_notifier_mapping.keys():
+                logging.warning(f"Cannot find mapping for {key} notifier.")
+            if self._config[key]["enable"]:
+                self._stats_reporter[key] = key_notifier_mapping[key](
+                    notify_manager=self._notify_manager, config=self._config[key]
+                )
+
+        if len(self._stats_reporter.values()) == 0:
+            logging.warning("No stats reporter enabled")
+        else:
+            # Start thread
+            self._is_running = True
+            self._thread = Thread(target=self._run_loop)
+            self._thread.start()
 
     def consume_harvester_messages(self, objects: List[HarvesterActivityMessage]):
-        if not self._enable:
-            return
-        for stat_acc in self._stat_accumulators:
-            if isinstance(stat_acc, HarvesterActivityConsumer):
-                for obj in objects:
-                    stat_acc.consume(obj)
+        for key in self._stats_reporter.keys():
+            self._stats_reporter[key].consume_harvester_messages(objects)
 
     def consume_signage_point_messages(self, objects: List[FinishedSignagePointMessage]):
-        if not self._enable:
-            return
-        for stat_acc in self._stat_accumulators:
-            if isinstance(stat_acc, FinishedSignageConsumer):
-                for obj in objects:
-                    stat_acc.consume(obj)
-
-    def _send_daily_notification(self):
-        summary = "Hello farmer! 👋 Here's what happened in the last 24 hours:\n"
-        for stat_acc in self._stat_accumulators:
-            summary += "\n" + stat_acc.get_summary()
-            stat_acc.reset()
-
-        self._notify_manager.process_events(
-            [Event(type=EventType.DAILY_STATS, priority=EventPriority.LOW, service=EventService.DAILY, message=summary)]
-        )
+        for key in self._stats_reporter.keys():
+            self._stats_reporter[key].consume_signage_point_messages(objects)
 
     def _run_loop(self):
         while self._is_running:
-            if datetime.now() > self._datetime_next_summary:
-                self._send_daily_notification()
-                self._datetime_next_summary += timedelta(days=1)
+            for key in self._stats_reporter.keys():
+                self._stats_reporter[key].report()
             sleep(1)
 
     def stop(self):
