+++ src/config.py
@@ -40,8 +40,8 @@ class Config:
     def get_keep_alive_monitor_config(self):
         return self._get_child_config("keep_alive_monitor", required=False)
 
-    def get_daily_stats_config(self):
-        return self._get_child_config("daily_stats")
+    def get_stats_config(self):
+        return self._get_child_config("stats")
 
 
 def check_keys(required_keys, config) -> bool:
